- Removed Classes and switched with Functional Components - Reduced some of the logic - Cleaned up some of the exports
