+++ frontend/workflows/projectCatalog/src/details/index.tsx
@@ -6,15 +6,16 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import GroupIcon from "@material-ui/icons/Group";
 import { capitalize } from "lodash";
 
-import type { DetailsCard, DetailWorkflowProps } from "..";
+import type { DetailWorkflowProps } from "..";
 
-import MetaCard from "./cards/meta";
 import type { ProjectInfo } from "./info/types";
 import fetchProject from "./resolvers/info";
-// import QuickLinksCard from "./quick-links";
+import type { DetailsCard } from "./card";
+import { MetaCard } from "./card";
 import { ProjectDetailsContext } from "./context";
 import ProjectHeader from "./header";
 import ProjectInfoCard from "./info";
+import QuickLinksCard from "./quick-links";
 
 type CardTyping = React.ReactNode | DetailsCard;
 
@@ -24,7 +25,7 @@ const StyledContainer = styled(Grid)({
 
 const DisabledItem = ({ name }: { name: string }) => (
   <Grid item>
-    <Tooltip title={`${name} is disabled`}>
+    <Tooltip title={`${capitalize(name)} is disabled`}>
       <FontAwesomeIcon icon={faLock} size="lg" />
     </Tooltip>
   </Grid>
@@ -63,60 +64,58 @@ const Details: React.FC<DetailWorkflowProps> = ({ children }) => {
   }, [children]);
 
   return (
-    <>
-      <ProjectDetailsContext.Provider value={{ projectInfo }}>
-        <StyledContainer container>
-          <Grid container direction="row">
-            <Grid container item xs={11}>
-              <Grid container direction="row" xs={12}>
-                <Grid item style={{ marginBottom: "22px" }}>
-                  {projectInfo && (
-                    <ProjectHeader name={projectInfo.name} description={projectInfo.description} />
-                  )}
-                </Grid>
-                <Grid container spacing={3}>
-                  <Grid container item direction="row" xs={4} spacing={2}>
-                    <Grid item xs={12}>
-                      <MetaCard
-                        title={capitalize(projectInfo?.name)}
-                        titleIcon={<GroupIcon />}
-                        fetchDataFn={() => fetchProject(projectId)}
-                        onSuccess={setProjectInfo}
-                        autoReload
-                        endAdornment={
-                          projectInfo?.disabled ? (
-                            <DisabledItem name={capitalize(projectInfo?.name)} />
-                          ) : null
-                        }
-                      >
-                        {projectInfo && <ProjectInfoCard data={projectInfo} />}
-                      </MetaCard>
-                    </Grid>
-                    {metaCards.length > 0 &&
-                      metaCards.map(card => (
-                        <Grid item xs={12}>
-                          {card}
-                        </Grid>
-                      ))}
-                  </Grid>
-                  <Grid container item direction="row" xs={8}>
-                    {dynamicCards.length > 0 &&
-                      dynamicCards.map(card => (
-                        <Grid item xs={12}>
-                          {card}
-                        </Grid>
-                      ))}
-                  </Grid>
-                </Grid>
-              </Grid>
-            </Grid>
-            {/* <Grid container item xs={1}>
-            <QuickLinksCard />
-          </Grid> */}
+    <ProjectDetailsContext.Provider value={{ projectInfo }}>
+      <StyledContainer container direction="row" wrap="nowrap">
+        {/* Column for project details and header */}
+        <Grid container item direction="column">
+          <Grid item style={{ marginBottom: "22px" }}>
+            {/* Static Header */}
+            {projectInfo && (
+              <ProjectHeader name={projectInfo.name} description={projectInfo.description} />
+            )}
           </Grid>
-        </StyledContainer>
-      </ProjectDetailsContext.Provider>
-    </>
+          <Grid container spacing={3}>
+            <Grid container item direction="row" xs={4} spacing={2}>
+              <Grid item xs={12}>
+                {/* Static Info Card */}
+                <MetaCard
+                  title={capitalize(projectInfo?.name)}
+                  titleIcon={<GroupIcon />}
+                  fetchDataFn={() => fetchProject(projectId)}
+                  onSuccess={setProjectInfo}
+                  autoReload
+                  endAdornment={
+                    projectInfo?.disabled ? <DisabledItem name={projectInfo?.name} /> : null
+                  }
+                >
+                  {projectInfo && <ProjectInfoCard data={projectInfo} />}
+                </MetaCard>
+              </Grid>
+              {/* Custom Meta Cards */}
+              {metaCards.length > 0 &&
+                metaCards.map(card => (
+                  <Grid item xs={12}>
+                    {card}
+                  </Grid>
+                ))}
+            </Grid>
+            <Grid container item direction="row" xs={8}>
+              {/* Custom Dynamic Cards */}
+              {dynamicCards.length > 0 &&
+                dynamicCards.map(card => (
+                  <Grid item xs={12}>
+                    {card}
+                  </Grid>
+                ))}
+            </Grid>
+          </Grid>
+        </Grid>
+        {/* Column for project quick links */}
+        <Grid container item direction="column" xs={1}>
+          <QuickLinksCard />
+        </Grid>
+      </StyledContainer>
+    </ProjectDetailsContext.Provider>
   );
 };
