+++ frontend/workflows/projectCatalog/src/details/quick-links.tsx
@@ -1,20 +1,27 @@
 import React from "react";
-import { Grid } from "@clutch-sh/core";
+import { Card, Grid } from "@clutch-sh/core";
 import { faGithub, faSlack } from "@fortawesome/free-brands-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-import { StyledCard } from "./cards/base";
-
 const QuickLinksCard = () => {
   return (
-    <StyledCard container direction="column" justify="space-evenly" alignItems="center">
-      <Grid item>
-        <FontAwesomeIcon icon={faGithub} size="3x" />
+    <Card>
+      <Grid
+        container
+        item
+        direction="column"
+        alignItems="center"
+        spacing={1}
+        style={{ padding: "7px 0" }}
+      >
+        <Grid item>
+          <FontAwesomeIcon icon={faGithub} size="3x" />
+        </Grid>
+        <Grid item>
+          <FontAwesomeIcon icon={faSlack} size="3x" />
+        </Grid>
       </Grid>
-      <Grid item>
-        <FontAwesomeIcon icon={faSlack} size="3x" />
-      </Grid>
-    </StyledCard>
+    </Card>
   );
 };
