+++ frontend/workflows/projectCatalog/src/index.tsx
@@ -1,17 +1,8 @@
 import type { BaseWorkflowProps, WorkflowConfiguration } from "@clutch-sh/core";
 
+import type { DetailsCard } from "./details/card";
 import Catalog from "./catalog";
 import Details from "./details";
-
-export { default as DynamicCard } from "./details/cards/dynamic";
-export { default as MetaCard } from "./details/cards/meta";
-
-export type DetailsCardTypes = "Dynamic" | "Metadata";
-
-export interface DetailsCard {
-  type: DetailsCardTypes;
-  title?: string;
-}
 
 export interface WorkflowProps extends BaseWorkflowProps {}
 export interface DetailWorkflowProps {
@@ -45,8 +36,10 @@ const register = (): WorkflowConfiguration => {
   };
 };
 
-export { LastEvent, LinkText, StyledLink, StyledRow } from "./details/cards/base";
+export { DynamicCard, MetaCard } from "./details/card";
+export { LastEvent } from "./details/helpers";
 export { useProjectDetailsContext } from "./details/context";
 export { Details as ProjectDetails };
+export type { DetailsCard };
 
 export default register;
